# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TGS Admin API Documentation",
        "description": "Documentation de l'API REST pour le système TGS Admin - Toulouse Game Show",
        "contact": {
            "email": "support@tgs.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur de développement"
        },
        {
            "url": "https://votre-domaine.com",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/api/v1/salons/{salon}/availabilities": {
            "get": {
                "tags": [
                    "Availabilities"
                ],
                "summary": "Liste des disponibilités",
                "description": "Récupère toutes les disponibilités d'un salon",
                "operationId": "getSalonAvailabilities",
                "parameters": [
                    {
                        "name": "salon",
                        "in": "path",
                        "description": "ID du salon",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Disponibilités récupérées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Vendredi"
                                                    },
                                                    "salon_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/salons/{salon}/availabilities/{availability}": {
            "get": {
                "tags": [
                    "Availabilities"
                ],
                "summary": "Détail d'une disponibilité",
                "description": "Récupère les informations d'une disponibilité spécifique",
                "operationId": "getSalonAvailability",
                "parameters": [
                    {
                        "name": "salon",
                        "in": "path",
                        "description": "ID du salon",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "availability",
                        "in": "path",
                        "description": "ID de la disponibilité",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Disponibilité récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "salon_id": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Disponibilité non trouvée"
                    }
                }
            }
        },
        "/api/v1/salons/{salon}/availabilities/{availability}/articles": {
            "get": {
                "tags": [
                    "Availabilities"
                ],
                "summary": "Articles d'une disponibilité",
                "description": "Récupère tous les articles avec une disponibilité spécifique",
                "operationId": "getAvailabilityArticles",
                "parameters": [
                    {
                        "name": "salon",
                        "in": "path",
                        "description": "ID du salon",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "availability",
                        "in": "path",
                        "description": "ID de la disponibilité",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Articles de la disponibilité récupérés",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Disponibilité non trouvée"
                    }
                }
            }
        },
...
    },
...
}